requirements.txt
Django==4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0